pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rf-lamp-mqtt-bridge"
version = "1.0.0"
description = "MQTT/RF Bridge for 433/315MHz Smart Lamps with HomeKit integration"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "paf0186", email = "paf0186@users.noreply.github.com"}
]
keywords = ["homekit", "mqtt", "rf", "433mhz", "smart-home", "raspberry-pi"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Home Automation",
    "Operating System :: POSIX :: Linux",
]

dependencies = [
    "paho-mqtt>=1.6.0",
    "RPi.GPIO>=0.7.0",
    "rpi-rf>=0.9.7",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]
dev = [
    "black>=22.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/paf0186/home_automation"
Repository = "https://github.com/paf0186/home_automation"
Issues = "https://github.com/paf0186/home_automation/issues"

[project.scripts]
rf-lamp-control = "lamp_control_mqtt:main"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.pytest_cache
  | __pycache__
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", ".pytest_cache"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["test_lamp_control.py"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=lamp_control_mqtt --cov-report=term-missing --cov-report=xml"

[tool.coverage.run]
source = ["lamp_control_mqtt"]
omit = ["test_*.py", "*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
